add CRD item PDO

+++ src/config/imagens.php
@@ -1,26 +1,29 @@
 <?php
-    function salvar($imagem){
-        $arquivo_tmp = $imagem['tmp_name'];
-        $nome_final = time().'.jpg';
-        $uploadfile = destiny().basename($nome_final);
+    class ImageRepository{
+        public static function salvar($image){
+            $nome_final = time().'.jpg';
+            $uploadfile = (self::destiny()).basename($nome_final);
 
-        if (move_uploaded_file($imagem['tmp_name'], $uploadfile)) {
-            return $nome_final;
-        } else {
-            throw new Exception("Não foi possível fazer o upload da imagem!");
+            if (move_uploaded_file($image['tmp_name'], $uploadfile)) {
+                return $nome_final;
+            } else {
+                throw new Exception("Não foi possível fazer o upload da imagem!");
+            }
         }
-    }
 
-    function destiny(){
-        return $_SERVER['DOCUMENT_ROOT']."/public/image/";
-    }
+        public static function destiny(){
+            return $_SERVER['DOCUMENT_ROOT']."/public/image/";
+        }
+    
+        public static function drop($image){
+            if(!self::exists($image)) return;
 
-    function drop($image){
-        $file = destiny().basename($image);
-        unlink($file);
-    }
-
-    function exists($image){
-        $file = destiny().basename($image);
-        return file_exists($file);
+            $file = (self::destiny()).basename($image);
+            unlink($file);
+        }
+    
+        public static function exists($image){
+            $file = (self::destiny()).basename($image);
+            return file_exists($file);
+        }
     }
